{{cookiecutter.project_slug}}/src/{{cookiecutter.project_slug}}/app/core/security.py
from datetime import datetime, timedelta, timezone

from jose import jwt
from passlib.hash import pbkdf2_sha256

from {{cookiecutter.project_slug}}.common.config import settings


def create_access_token(data: dict, expires_delta: timedelta = timedelta(minutes=30)) -> str:
    to_encode = data.copy()
    expire = datetime.now(timezone.utc) + expires_delta
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, settings.SECRET_KEY, algorithm="HS256")
    return encoded_jwt


def verify_password(plain_password: str, hashed_password: str) -> bool:
    return pbkdf2_sha256.verify(plain_password, hashed_password)


def get_password_hash(password: str) -> str:
    return pbkdf2_sha256.hash(password)
